Moved `Index` and related tests to submodule `people::index`.
